insert, delete を配列で指定できるようにして一括処理するように変更

+++ node/src/dao/YokohamaLibraryDao.ts
@@ -1,6 +1,11 @@
 import * as sqlite from 'sqlite';
 import sqlite3 from 'sqlite3';
 
+export interface Book {
+	type: string;
+	title: string;
+}
+
 /**
  * 横浜市立図書館
  */
@@ -42,48 +47,11 @@ export default class YokohamaLibraryDao {
 	}
 
 	/**
-	 * 受取可データを取得する
+	 * すべての受取可能データを取得する
 	 *
-	 * @param data - 検索データ
-	 *
-	 * @returns 登録データ
+	 * @returns 受取可能データ
 	 */
-	async selectAvailable(data: YokohamaLibraryDb.Available): Promise<YokohamaLibraryDb.Available | null> {
-		const dbh = await this.getDbh();
-
-		const sth = await dbh.prepare(`
-			SELECT
-				type,
-				title
-			FROM
-				d_available
-			WHERE
-				type = :type AND
-				title = :title
-		`);
-		await sth.bind({
-			':type': data.type,
-			':title': data.title,
-		});
-		const row = await sth.get<YokohamaLibraryDb.Available>();
-		await sth.finalize();
-
-		if (row === undefined) {
-			return null;
-		}
-
-		return {
-			type: row.type,
-			title: row.title,
-		};
-	}
-
-	/**
-	 * 受取可データを全取得する
-	 *
-	 * @returns 登録データ
-	 */
-	async selectAvailables(): Promise<YokohamaLibraryDb.Available[]> {
+	async selectAvailables(): Promise<Book[]> {
 		const dbh = await this.getDbh();
 
 		const sth = await dbh.prepare(`
@@ -96,39 +64,84 @@ export default class YokohamaLibraryDao {
 		const rows = await sth.all<YokohamaLibraryDb.Available[]>();
 		await sth.finalize();
 
-		const data: YokohamaLibraryDb.Available[] = [];
+		const datas: Book[] = [];
 		for (const row of rows) {
-			data.push({
+			datas.push({
 				type: row.type,
 				title: row.title,
 			});
 		}
 
-		return data;
+		return datas;
 	}
 
 	/**
-	 * 受取可データを登録する
+	 * 指定されたデータが登録済みかどうかチェックする
 	 *
-	 * @param data - 登録データ
+	 * @param data - 書籍データ
+	 *
+	 * @returns 登録済みなら true
 	 */
-	async insertAvailable(data: YokohamaLibraryDb.Available): Promise<void> {
+	async isRegisted(data: Book): Promise<boolean> {
+		interface Select {
+			count: number;
+		}
+
+		const dbh = await this.getDbh();
+
+		const sth = await dbh.prepare(`
+			SELECT
+				COUNT(title) AS count
+			FROM
+				d_available
+			WHERE
+				type = :type AND
+				title = :title
+		`);
+		await sth.bind({
+			':type': data.type,
+			':title': data.title,
+		});
+		const row = await sth.get<Select>();
+		await sth.finalize();
+
+		if (row === undefined) {
+			return false;
+		}
+
+		return row.count > 0;
+	}
+
+	/**
+	 * 受取可能データを登録する
+	 *
+	 * @param datas - 登録するデータ
+	 */
+	async insertAvailable(datas: Book[]): Promise<void> {
+		if (datas.length === 0) {
+			return;
+		}
+
 		const dbh = await this.getDbh();
 
 		await dbh.exec('BEGIN');
 		try {
-			const sth = await dbh.prepare(`
-				INSERT INTO
-					d_available
-					(type, title)
-				VALUES
-					(:type, :title)
-			`);
-			await sth.run({
-				':type': data.type,
-				':title': data.title,
-			});
-			await sth.finalize();
+			await Promise.all(
+				datas.map(async (data) => {
+					const sth = await dbh.prepare(`
+						INSERT INTO
+							d_available
+							(type, title)
+						VALUES
+							(:type, :title)
+					`);
+					await sth.run({
+						':type': data.type,
+						':title': data.title,
+					});
+					await sth.finalize();
+				}),
+			);
 
 			await dbh.exec('COMMIT');
 		} catch (e) {
@@ -138,27 +151,35 @@ export default class YokohamaLibraryDao {
 	}
 
 	/**
-	 * 受取可データを削除する
+	 * 受取可能データを削除する
 	 *
-	 * @param data - 削除データ
+	 * @param datas - 削除するデータ
 	 */
-	async deleteAvailable(data: YokohamaLibraryDb.Available): Promise<void> {
+	async deleteAvailable(datas: Book[]): Promise<void> {
+		if (datas.length === 0) {
+			return;
+		}
+
 		const dbh = await this.getDbh();
 
 		await dbh.exec('BEGIN');
 		try {
-			const sth = await dbh.prepare(`
-				DELETE FROM
-					d_available
-				WHERE
-					type = :type AND
-					title = :title
-			`);
-			await sth.run({
-				':type': data.type,
-				':title': data.title,
-			});
-			await sth.finalize();
+			await Promise.all(
+				datas.map(async (data) => {
+					const sth = await dbh.prepare(`
+						DELETE FROM
+							d_available
+						WHERE
+							type = :type AND
+							title = :title
+					`);
+					await sth.run({
+						':type': data.type,
+						':title': data.title,
+					});
+					await sth.finalize();
+				}),
+			);
 
 			await dbh.exec('COMMIT');
 		} catch (e) {
